Refactoring of parsing scripts+ shape_indices
The JSON script now also writes a JSON file called shape_indices.json, which given a route and a shape (i.e. GPS coordinate), yields the index of that particular coordinate in the route sequence.

This allows the app to determine the exact subsequence of points between two stops (since we know their GPS coordinates and can perform a simple slice through the array using the indices), which we can pass to the client to render via Mapbox; this simplifies things, since each rendering can simply treat this subsequence as the only path of concern when moving the train.

Note that this assumes that each route's GPS coordinates are unique within that route -- i.e., no route contains a cycle.